Refactor game structure by introducing GameManager and ConsoleGame classes for improved game flow and logic management. Update Program.cs to reflect new namespaces and enhance README with file locations for better navigation.
